requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
anime-api==1.0.2
aniwrap==1.0.1
attrs==23.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
Deprecated==1.2.14
discord==2.3.2
discord.py==2.3.2
frozenlist==1.4.1
idna==3.6
jaconv==0.3.4
jikanpy-v4==1.0.2
multidict==6.0.4
pykakasi==2.2.1
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
setuptools==69.0.3
simplejson==3.19.2
six==1.16.0
urllib3==2.1.0
wrapt==1.16.0
yarl==1.9.4
